feat: 126 authorization - gates

+++ src/app/Http/Controllers/PostController.php
@@ -38,6 +38,7 @@ class PostController extends Controller
      */
     public function create()
     {
+        $this->authorize('create_post');
         $categories = Category::all();
         return view('create', compact('categories'));
     }
@@ -47,6 +48,8 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create_post');
+
         $request->validate([
             'image' => ['required', 'max:2028', 'image'],
             'title' => ['required', 'max: 255'],
@@ -84,6 +87,8 @@ class PostController extends Controller
      */
     public function edit(string $id)
     {
+        $this->authorize('edit_post');
+
         $post = Post::findOrFail($id);
         $categories = Category::all();
         return view('edit', compact('post', 'categories'));
@@ -94,6 +99,7 @@ class PostController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize('edit_post');
         
         $request->validate([
             'title' => ['required', 'max: 255'],
@@ -130,6 +136,8 @@ class PostController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('delete_post');
+
         $post = Post::findOrFail($id);
         $post->delete();
 
@@ -138,12 +146,16 @@ class PostController extends Controller
 
     public function trashed()
     {
+        $this->authorize('delete_post');
+
         $posts = Post::onlyTrashed()->get();
         return view('trashed', compact('posts'));
     }
 
     public function restore($id)
     {
+        $this->authorize('delete_post');
+
         $post = Post::onlyTrashed()->findOrFail($id);
         $post->restore();
 
@@ -152,6 +164,8 @@ class PostController extends Controller
 
     public function forceDelete($id)
     {
+        $this->authorize('delete_post');
+        
         $post = Post::onlyTrashed()->findOrFail($id);
 
         File::delete(public_path($post->image));
